Move common objects into separate file to remove duplication
