+++ grafana-plugin/src/pages/integration_2/Integration2.helper.ts
@@ -39,24 +39,6 @@ const IntegrationHelper = {
     return slice.length === line.length ? slice : `${slice} ...`;
   },
 
-  getRouteConditionWording(channelFilters: Array<ChannelFilter['id']>, routeIndex: number): 'Default' | 'Else' | 'If' {
-    const totalCount = Object.keys(channelFilters).length;
-
-    if (routeIndex === totalCount - 1) {
-      return 'Default';
-    }
-    return routeIndex ? 'Else' : 'If';
-  },
-
-  getRouteConditionTooltipWording(channelFilters: Array<ChannelFilter['id']>, routeIndex: number) {
-    const totalCount = Object.keys(channelFilters).length;
-
-    if (routeIndex === totalCount - 1) {
-      return 'If the alert payload does not match to the previous routes, it will be directed to this default route.';
-    }
-    return 'If the alert payload evaluates the route template as True, it will be directed to this route. It will not be evaluated against the subsequent routes.';
-  },
-
   getMaintenanceText(maintenanceUntill: number, mode: number = undefined) {
     const date = dayjs(new Date(maintenanceUntill * 1000));
     const now = dayjs();
